requirements.txt
# Core dependencies
playwright==1.40.0
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/LLM for message generation (optional - can use local models)
openai==1.3.0
anthropic==0.7.0

# Resume parsing
python-docx==1.1.0
PyPDF2==3.0.1

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
faker==20.1.0

# Logging and monitoring
loguru==0.7.2
python-dateutil==2.8.2

# Database for tracking (SQLite)
sqlalchemy==2.0.23

# Configuration
pyyaml==6.0.1

